My ACM\Topcoder algo library consolidated.
This is the first archive bulk committ for myriad files maintained since early 2003.
None of the code here is published in any solution \project format though, but is intended
to be used holistically with detailed testing in personal projects.

WARNING: The repo is still looking @ me its with eyes-wide-open, hungry for updates.
DISCLAIMER: None of the code is viral w.r.t nature , but watch out for unforeseen
segfaults and buffer overruns due to inefficient or so data structures.
So, use the code at your own risk... ;)